updateWarehouse partially functional
removed warehouse from WarehouseName toString() to prevent infinite loop
added findWarehouseById Service method with exception handling
added updateWarehouse Service method with exception handling
added updateWarehouse Controller endpoint

+++ Back-End/src/main/java/com/mtlaguerre/inventory_management/repositories/WarehouseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.mtlaguerre.inventory_management.models.Warehouse;
 
 @Repository
-public interface WarehouseRepository extends JpaRepository<Warehouse, Integer>{
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long>{
 
 }
